Fix Codeclimate issue
