+++ Tunny/Util/PythonInstaller.cs
@@ -71,40 +71,18 @@ namespace Tunny.Util
         {
             return new[]
             {
-                "alembic",
-                "attrs",
-                "autopage",
-                "bottle",
-                "cliff",
-                "cmaes",
-                "cmd2",
-                "colorama",
-                "colorlog",
-                "greenlet",
-                "joblib",
-                "Mako",
-                "MarkupSafe",
-                "numpy",
-                "optuna",
-                "optuna-dashboard",
-                "packaging",
-                "pbr",
-                "plotly",
-                "prettytable",
-                "pyparsing",
-                "pyperclip",
-                "pyreadline3",
-                "PyYAML",
-                "scikit-learn",
-                "scipy",
-                "six",
-                "sklearn",
-                "SQLAlchemy",
-                "stevedore",
-                "tenacity",
-                "threadpoolctl",
-                "tqdm",
-                "wcwidth",
+                "alembic",    "attrs",    "autopage",
+                "bottle",     "cliff",    "cmaes",
+                "cmd2",       "colorama", "colorlog",
+                "greenlet",   "joblib",   "Mako",
+                "MarkupSafe", "numpy",    "optuna",
+                "optuna-dashboard",       "packaging",
+                "pbr",        "plotly",   "prettytable",
+                "pyparsing",  "pyperclip","pyreadline3",
+                "PyYAML",     "scikit-learn",
+                "scipy",      "six",      "sklearn",
+                "SQLAlchemy", "stevedore","tenacity",
+                "threadpoolctl", "tqdm", "wcwidth",
             };
         }
     }
